Move `to_bytes` and `to_unicode` from `utils` to `strings`
