+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This project is in alpha. I plan to adhere to [semantic versioning][semver]
 once the API is stable.
 
 ## Unreleased
+- CHANGED: Moved `to_bytes` from `framewirc.utils` to `framewirc.strings`.
+
+- CHANGED: Moved `to_unicode` from `framewirc.utils` to `framewirc.strings`.
+
 - CHANGED: Moved `chunk_message` from `framewirc.utils` to `framewirc.message`.
 
 - REMOVED: Support for Python `3.4` has been removed.
